consolidate os files
