+++ README.rst
@@ -10,6 +10,10 @@ hpsspy
     :target: https://coveralls.io/github/weaverba137/hpsspy?branch=pre-0.3
     :alt: Test Coverage Status
 
+.. image:: https://readthedocs.org/projects/hpsspy/badge/?version=latest
+    :target: http://hpsspy.readthedocs.io/en/latest/?badge=latest
+    :alt: Documentation Status
+
 Overview
 --------
